Gestion des exceptions + corrections de bugs avec des static_cast pour les .size() + plus d'iterateur inutile pour getHexagone de Tuile

+++ Code/chantier.h
@@ -13,7 +13,7 @@ public:
     // Gestion des tuiles
     void ajouterPile(Pile& p);
 
-    size_t getNbTuiles() const { return tuiles.size(); }
+    int getNbTuiles() const { return static_cast<int>(tuiles.size()); }
 
     void retirerTuile(Tuile* t); // Modifié pour prendre un pointeur
     bool estVide() const;
